+++ public/locales/en.json
@@ -73,6 +73,7 @@
     "takesXFrom": "Takes {{ x }} from",
     "chooseAbility": "Choose ability",
     "chooseTeraType": "Choose tera type",
+    "permalink": "Copy link",
     "abilityNames": {
       "none": "None",
       "filter": "Filter",
