+++ src/App.module.css
@@ -9,9 +9,9 @@
   padding: 3rem;
   border-radius: 9999px;
   background: radial-gradient(hsl(0 0% 100% / 40%), hsl(0 0% 100% / 0%));
-  backdrop-filter: blur(8px);
+  backdrop-filter: blur(4px);
   box-shadow: inset 0 0 16px hsl(0 0% 100% / 50%), 0 0 0 1px hsl(0 0% 0% / 10%),
-    0 4px 8px hsl(0 0% 0% / 10%), 0 0 0 200vmax hsl(0 0% 0% / 20%);
+    0 4px 8px hsl(0 0% 0% / 10%), 0 0 16px hsl(0 0% 0% / 20%);
 }
 
 .easterEgg[data-animate="true"] {
@@ -29,12 +29,7 @@
 }
 
 .easterEgg img {
-  filter: drop-shadow(0 0 1px hsl(0 0% 100% / 20%))
-    drop-shadow(0 0 2px hsl(0 0% 100% / 20%))
-    drop-shadow(0 0 3px hsl(0 0% 100% / 20%))
-    drop-shadow(0 0 4px hsl(0 0% 100% / 20%))
-    drop-shadow(0 0 5px hsl(0 0% 100% / 20%))
-    drop-shadow(0 0 6px hsl(0 0% 100% / 20%));
+  filter: drop-shadow(0 0 16px hsl(0 0% 100%));
 }
 
 @keyframes easterEgg_animation {
